+++ bin/clean.js
@@ -1,8 +1,10 @@
 #!/usr/bin/env node
 // This lets us remove files on all platforms. Notably, `rm` is missing on Windows.
-const path = require("path");
-const fs = require("fs");
+import * as path from "path";
+import * as fs from "fs";
+import { fileURLToPath } from "url";
 
-const distPath = path.join(__dirname, "..", "dist");
+const thisPath = fileURLToPath(import.meta.url);
+const distPath = path.join(path.dirname(thisPath), "..", "dist");
 
 fs.rmSync(distPath, { recursive: true, force: true });