+++ middlewares/origin-agent-cluster/index.ts
@@ -11,5 +11,4 @@ function originAgentCluster() {
   };
 }
 
-module.exports = originAgentCluster;
 export default originAgentCluster;